requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.4
tzdata==2023.3
psycopg[binary,pool]==3.1.8
python-decouple